[wip][win32] Port NBD daemon
We're planning to add a Windows virtual storport miniport driver
that's going to allow mounting rbd images. In fact, we're going
to provide an NBD compatible interface that rbd-nbd (and other
projects, e.g. qemu-nbd) may use.

Instead of an unix socket, tcp sockets as well as a shared memory
mechanism may be used as an nbd communication channel.

The functions that initialize the device will have to be abstracted.
This commit provides just some stubs that will have to be filled in.